Fix all tests. Now passing with limited data and assertions

+++ geodata-repository/src/main/java/com/ethlo/geodata/model/RTreePayload.java
@@ -10,12 +10,12 @@ package com.ethlo.geodata.model;
  * it under the terms of the GNU Lesser General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Lesser Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Lesser Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/lgpl-3.0.html>.
@@ -26,18 +26,18 @@ import org.locationtech.jts.geom.Envelope;
 
 public class RTreePayload
 {
-    private final long id;
+    private final int id;
     private final double area;
     private final Envelope envelope;
-    
-    public RTreePayload(long id, double area, Envelope envelope)
+
+    public RTreePayload(int id, double area, Envelope envelope)
     {
         this.id = id;
         this.area = area;
         this.envelope = envelope;
     }
 
-    public long getId()
+    public int getId()
     {
         return id;
     }
